feat: move matomo tracking function to utils
